+++ src/main/scala/com/github/everpeace/k8s/throttler/controller/ThrottleController.scala
@@ -26,6 +26,11 @@ import com.github.everpeace.k8s.throttler.KubeThrottleConfig
 import com.github.everpeace.k8s.throttler.controller.ThrottleController._
 import com.github.everpeace.k8s.throttler.crd.v1alpha1
 import com.github.everpeace.k8s.throttler.crd.v1alpha1.Implicits._
+import com.github.everpeace.k8s.throttler.metrics.{
+  ClusterThrottleControllerMetrics,
+  MetricsBase,
+  ThrottleControllerMetrics
+}
 import kamon.Kamon
 import play.api.libs.json.Format
 import skuber.Resource.ResourceList
@@ -556,241 +561,6 @@ object ThrottleController {
     implicit val _config = config
     Props(new ThrottleController())
   }
-}
-
-trait MetricsBase {
-  def resourceQuantityToLong(q: (String, Resource.Quantity)): Long = q._1 match {
-    case Resource.cpu =>
-      (q._2.amount * 1000).toLong // convert to milli
-    case _ => q._2.amount.toLong
-  }
-
-  def resourceQuantityToTag(q: (String, _)): kamon.Tags = Map("resource" -> q._1)
-}
-
-trait ClusterThrottleControllerMetrics extends MetricsBase {
-  self: {
-    def log: {
-      def info(s: String): Unit
-      def debug(s: String): Unit
-    }
-  } =>
-
-  def resetClusterThrottleMetric(clthrottle: v1alpha1.ClusterThrottle): Unit = {
-    val zeroSpec = clthrottle.spec.copy(
-      threshold = clthrottle.spec.threshold.copy(
-        podsCount = Option(0),
-        resourceRequests =
-          clthrottle.spec.threshold.resourceRequests.mapValues(_ => Resource.Quantity("0"))
-      ),
-      selector = clthrottle.spec.selector
-    )
-
-    val zeroFalseStatus = clthrottle.status.map(
-      st =>
-        st.copy(
-          throttled = st.throttled.copy(
-            podsCount = st.throttled.podsCount.map(_ => false),
-            resourceRequests = st.throttled.resourceRequests.mapValues(_ => false)
-          ),
-          used = st.used.copy(
-            podsCount = st.used.podsCount.map(_ => 0),
-            resourceRequests = st.used.resourceRequests.mapValues(_ => Resource.Quantity("0"))
-          )
-      )
-    )
-    val zero = clthrottle.copy(
-      spec = zeroSpec,
-      status = zeroFalseStatus
-    )
-
-    recordClusterThrottleSpecMetric(zero)
-    recordClusterThrottleStatusMetric(zero)
-  }
-
-  def recordClusterThrottleSpecMetric(clthrottle: v1alpha1.ClusterThrottle): Unit = {
-    val metadataTags = metadataTagsForClusterThrottle(clthrottle)
-
-    val resourceRequestsThresholdGauge =
-      Kamon.gauge("clusterthrottle.spec.threshold.resourceRequests")
-    val podsCountsThresholdGauge = Kamon.gauge("clusterthrottle.spec.threshold.podsCount")
-
-    clthrottle.spec.threshold.resourceRequests.foreach { rq =>
-      val tags  = metadataTags ++ resourceQuantityToTag(rq)
-      val value = resourceQuantityToLong(rq)
-      log.info(s"setting gauge '${resourceRequestsThresholdGauge.name}{${tags.values
-        .mkString(",")}}' value with ${value}")
-      resourceRequestsThresholdGauge.refine(tags).set(value)
-    }
-
-    clthrottle.spec.threshold.podsCount.foreach { value =>
-      val tags = metadataTags
-      log.info(
-        s"setting gauge '${podsCountsThresholdGauge.name}{${tags.values.mkString(",")}}' value with ${value}")
-      podsCountsThresholdGauge.refine(tags).set(value)
-    }
-  }
-
-  def recordClusterThrottleStatusMetric(clthrottle: v1alpha1.ClusterThrottle): Unit = {
-    val metadataTags  = metadataTagsForClusterThrottle(clthrottle)
-    val statusRRGauge = Kamon.gauge("clusterthrottle.status.throttled.resourceRequests")
-    val statusPCGauge = Kamon.gauge("clusterthrottle.status.throttled.podsCount")
-    val usedRRGauge   = Kamon.gauge("clusterthrottle.status.used.resourceRequests")
-    val usedPCGauge   = Kamon.gauge("clusterthrottle.status.used.podsCount")
-
-    val b2i = (b: Boolean) => b compare false
-    clthrottle.status.foreach { status =>
-      status.throttled.resourceRequests foreach { rq =>
-        val tags = metadataTags ++ resourceQuantityToTag(rq)
-        log.info(
-          s"setting gauge '${statusRRGauge.name}{${tags.values.mkString(",")}}' value with ${b2i(rq._2)}")
-        statusRRGauge.refine(tags).set(b2i(rq._2))
-      }
-      status.throttled.podsCount foreach { value =>
-        val tags = metadataTags
-        log.info(
-          s"setting gauge '${statusPCGauge.name}{${tags.values.mkString(",")}}' value with ${b2i(value)}")
-        statusPCGauge.refine(tags).set(b2i(value))
-
-      }
-
-      status.used.resourceRequests foreach { rq =>
-        val tags  = metadataTags ++ resourceQuantityToTag(rq)
-        val value = resourceQuantityToLong(rq)
-        log.debug(
-          s"setting gauge '${usedRRGauge.name}{${tags.values.mkString(",")}}' value with ${value}")
-        usedRRGauge.refine(tags).set(value)
-      }
-      status.used.podsCount foreach { value =>
-        val tags = metadataTags
-        log.debug(
-          s"setting gauge '${usedPCGauge.name}{${tags.values.mkString(",")}}' value with ${value}")
-        usedPCGauge.refine(tags).set(value)
-      }
-    }
-  }
-
-  def metadataTagsForClusterThrottle(throttle: v1alpha1.ClusterThrottle): kamon.Tags = {
-    val metadata = List(
-      "name" -> throttle.metadata.name,
-      "uuid" -> throttle.metadata.uid
-    ).toMap ++ throttle.metadata.clusterName.map(cn => Map("cluster" -> cn)).getOrElse(Map.empty)
-
-    metadata ++ throttle.metadata.labels ++ throttle.metadata.annotations
-  }
-
-}
-
-trait ThrottleControllerMetrics extends MetricsBase {
-  self: {
-    def log: {
-      def info(s: String): Unit
-      def debug(s: String): Unit
-    }
-  } =>
-
-  def resetThrottleMetric(throttle: v1alpha1.Throttle): Unit = {
-    val zeroSpec = throttle.spec.copy(
-      threshold = throttle.spec.threshold.copy(
-        podsCount = Option(0),
-        resourceRequests =
-          throttle.spec.threshold.resourceRequests.mapValues(_ => Resource.Quantity("0"))
-      ),
-      selector = throttle.spec.selector
-    )
-
-    val zeroFalseStatus = throttle.status.map(
-      st =>
-        st.copy(
-          throttled = st.throttled.copy(
-            podsCount = st.throttled.podsCount.map(_ => false),
-            resourceRequests = st.throttled.resourceRequests.mapValues(_ => false)
-          ),
-          used = st.used.copy(
-            podsCount = st.used.podsCount.map(_ => 0),
-            resourceRequests = st.used.resourceRequests.mapValues(_ => Resource.Quantity("0"))
-          )
-      ))
-    val zero = throttle.copy(
-      spec = zeroSpec,
-      status = zeroFalseStatus
-    )
-
-    recordThrottleSpecMetric(zero)
-    recordThrottleStatusMetric(zero)
-  }
-
-  def recordThrottleSpecMetric(throttle: v1alpha1.Throttle): Unit = {
-    val metadataTags = metadataTagsForThrottle(throttle)
-
-    val resourceRequestsThresholdGauge = Kamon.gauge("throttle.spec.threshold.resourceRequests")
-    val podsCountsThresholdGauge       = Kamon.gauge("throttle.spec.threshold.podsCount")
-
-    throttle.spec.threshold.resourceRequests.foreach { rq =>
-      val tags  = metadataTags ++ resourceQuantityToTag(rq)
-      val value = resourceQuantityToLong(rq)
-      log.info(s"setting gauge '${resourceRequestsThresholdGauge.name}{${tags.values
-        .mkString(",")}}' value with ${value}")
-      resourceRequestsThresholdGauge.refine(tags).set(value)
-    }
-
-    throttle.spec.threshold.podsCount.foreach { value =>
-      val tags = metadataTags
-      log.info(
-        s"setting gauge '${podsCountsThresholdGauge.name}{${tags.values.mkString(",")}}' value with ${value}")
-      podsCountsThresholdGauge.refine(tags).set(value)
-    }
-  }
-
-  def recordThrottleStatusMetric(throttle: v1alpha1.Throttle): Unit = {
-    val metadataTags  = metadataTagsForThrottle(throttle)
-    val statusRRGauge = Kamon.gauge("throttle.status.throttled.resourceRequests")
-    val statusPCGauge = Kamon.gauge("throttle.status.throttled.podsCount")
-    val usedRRGauge   = Kamon.gauge("throttle.status.used.resourceRequests")
-    val usedPCGauge   = Kamon.gauge("throttle.status.used.podsCount")
-
-    val b2i = (b: Boolean) => b compare false
-    throttle.status.foreach { status =>
-      status.throttled.resourceRequests foreach { rq =>
-        val tags = metadataTags ++ resourceQuantityToTag(rq)
-        log.info(
-          s"setting gauge '${statusRRGauge.name}{${tags.values.mkString(",")}}' value with ${b2i(rq._2)}")
-        statusRRGauge.refine(tags).set(b2i(rq._2))
-      }
-      status.throttled.podsCount foreach { value =>
-        val tags = metadataTags
-        log.info(
-          s"setting gauge '${statusPCGauge.name}{${tags.values.mkString(",")}}' value with ${b2i(value)}")
-        statusPCGauge.refine(tags).set(b2i(value))
-
-      }
-
-      status.used.resourceRequests foreach { rq =>
-        val tags  = metadataTags ++ resourceQuantityToTag(rq)
-        val value = resourceQuantityToLong(rq)
-        log.debug(
-          s"setting gauge '${usedRRGauge.name}{${tags.values.mkString(",")}}' value with ${value}")
-        usedRRGauge.refine(tags).set(value)
-      }
-      status.used.podsCount foreach { value =>
-        val tags = metadataTags
-        log.debug(
-          s"setting gauge '${usedPCGauge.name}{${tags.values.mkString(",")}}' value with ${value}")
-        usedPCGauge.refine(tags).set(value)
-      }
-    }
-  }
-
-  def metadataTagsForThrottle(throttle: v1alpha1.Throttle): kamon.Tags = {
-    val metadata = List(
-      "name"      -> throttle.metadata.name,
-      "namespace" -> throttle.metadata.namespace,
-      "uuid"      -> throttle.metadata.uid
-    ).toMap ++ throttle.metadata.clusterName.map(cn => Map("cluster" -> cn)).getOrElse(Map.empty)
-
-    metadata ++ throttle.metadata.labels ++ throttle.metadata.annotations
-  }
-
 }
 
 trait ClusterThrottleControllerLogic {
